// api/src/common/enums.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  SUPERVISOR = 'SUPERVISOR',
  ANALYST = 'ANALYST',
  OPERATOR = 'OPERATOR',
}

export enum InspectionStatus {
  ENVIADO = 'ENVIADO',
  ABERTO = 'ABERTO',
  AGUARDANDO_COLETA = 'AGUARDANDO_COLETA',
  EM_ANDAMENTO = 'EM_ANDAMENTO',
  NOVA_COLETA = 'NOVA_COLETA',
  NOVA_COLETA_FINALIZADA = 'NOVA_COLETA_FINALIZADA',
  APROVADO = 'APROVADO',
  APROVADO_COM_RESSALVAS = 'APROVADO_COM_RESSALVAS',
  REPROVADO = 'REPROVADO',
  FINALIZADO = 'FINALIZADO',
}


export enum AnalysisStatus {
  EM_ANDAMENTO = 'EM_ANDAMENTO',
  FINALIZADO = 'FINALIZADO',
  APROVADO_COM_RESSALVA = 'APROVADO_COM_RESSALVA',
  REPROVADO = 'REPROVADO',
  NOVA_COLETA = 'NOVA_COLETA',
  NOVA_COLETA_FINALIZADA = 'NOVA_COLETA_FINALIZADA',
}